Aggregate player statistics from a particular division and season

This is a wrapper around group_player_stats() that extracts all season data from players in all teams from the chosen division. It aggregates all player data and team data into separate data frames and combines them into a list.

Usage

division_player_stats(
  year = NULL,
  division = 1,
  sport = "WVB",
  save = FALSE,
  path = ""
)

Arguments

year

Year for fall of desired season.

division

NCAA division (must be 1, 2, or 3).

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

save

Logical for whether to save the statistics locally as CSVs (default FALSE).

path

Character string of path to save statistics files.

Value

Returns list with player statistics and team statistics data frames.

See also

Examples

if (FALSE) { # \dontrun{
division_player_stats(2024)
division_player_stats(2023, division = 3, sport = "MVB")
division_player_stats(2024, save = TRUE, path = "data/")
} # }