Extract data frame of team names, IDs, conference, division, and season

NCAA datasets use a unique ID for each sport, team, and season. This function extracts team names, IDs, and conferences for each NCAA team in a division. However, you should not need to use this function for volleyball data from 2020-2024, as it has been used to generate wvb_teams and mvb_teams. However, it is available to use for other sports, using the appropriate three letter sport code drawn from ncaa_sports (e.g., men's baseball is "MBA").

Usage

get_teams(year = NULL, division = 1, sport = "WVB")

Arguments

year

Single numeric year for fall of desired season.

division

NCAA division (must be 1, 2, or 3).

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

Value

Returns a data frame of all teams, their team ID, division, conference, and season.

Note

This function is a modification of the ncaa_teams() function from the {baseballr} package.

Examples

if (FALSE) { # \dontrun{
get_teams(year = 2024)
get_teams(year = 2023, division = 3, sport = "MVB")
} # }