Aggregate player statistics from a set of teams for a particular season

This function runs player_season_stats() on a vector of team names to extract season data for the players and teams in the vector. It aggregates all player data and team data into separate data frames and combines them into a list. For instance, if you want to extract the data from the teams in the women's 2024 Final Four, pass a vector of c("Louisville", "Nebraska", "Penn State", "Pittsburgh") to the function. Team names can be found in ncaa_teams or by using find_team_name().

Usage

group_player_stats(teams = NULL, year = NULL, sport = "WVB")

Arguments

teams

Character vector of team names to aggregate.

year

Year for fall of desired season.

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

Value

Returns list with player statistics and team statistics data frames.

See also

Examples

if (FALSE) { # \dontrun{
group_player_stats(teams = c("Louisville", "Nebraska", "Penn St.", "Pittsburgh"),
year = 2024)
group_player_stats(teams = c("UCLA", "Long Beach St."), year = 2023, sport = "MVB")
} # }