Extract player statistics from a particular team and season

The NCAA's main page for a team includes a tab called "Team Statistics". This function extracts the table of player statistics for the season, as well as team and opponent statistics (though these can be omitted).

Usage

team_player_stats(team_id, team_stats = TRUE)

Arguments

team_id

Team ID determined by NCAA for season. To find ID, use find_team_id().

team_stats

Logical indicating whether to include (TRUE) or exclude (FALSE) team statistics. Default includes team statistics with player statistics.

Value

Returns a data frame of player statistics. Note that hometown and high school were added in 2024.

See also

Examples

if (FALSE) { # \dontrun{
team_player_stats("585290")
team_player_stats(find_team_id("Nebraska", 2024), team_stats = FALSE)
team_player_stats(find_team_id("UCLA", 2023, sport = "MVB"))
} # }