Aggregate player statistics for a NCAA conference and seasons

This is a wrapper around group_stats() that extracts season, match, or pbp data from players in all teams in the chosen conference. For season stats, it aggregates all player data and team data into separate data frames and combines them into a list. For match and pbp stats, it aggregates into a data frame. Conferences names can be found in ncaa_conferences.

Usage

conference_stats(
  year = NULL,
  conf = NULL,
  level = NULL,
  sport = "WVB",
  save = FALSE,
  path = "."
)

Arguments

year

Numeric vector of years for fall of desired seasons.

conf

NCAA conference name.

level

Character string defining whether to aggregate "season", "match", or play-by-play ("pbp") data.

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

save

Logical for whether to save the statistics locally as CSVs (default FALSE).

path

Character string of path to save statistics files.

Value

For season level, returns list with data frames of player statistics and team statistics. For match and pbp levels, returns data frame of player statistics and play-by-play information respectively.

See also

Examples

# \donttest{
conference_stats(year = 2024, conf = "Peach Belt", level = "season")
#> $playerdata
#> # A tibble: 81 × 29
#>    Season    Team    Conference Number Player         Yr    Pos   Ht    Hometown
#>    <chr>     <chr>   <chr>       <dbl> <chr>          <chr> <chr> <chr> <chr>   
#>  1 2024-2025 Augusta Peach Belt      2 Marah Zenner   Sr    S     5-5   ""      
#>  2 2024-2025 Augusta Peach Belt      3 Rylie Harris   Sr    L/DS  5-5   ""      
#>  3 2024-2025 Augusta Peach Belt      4 Giulia Rodrig… Sr    DS    5-3   "Boca R…
#>  4 2024-2025 Augusta Peach Belt      5 Mia Saavedra   So    S     5-10  ""      
#>  5 2024-2025 Augusta Peach Belt      7 Mikayla Kline  So    OH    5-10  ""      
#>  6 2024-2025 Augusta Peach Belt      8 Sophia Kruczko Sr    RS    6-3   ""      
#>  7 2024-2025 Augusta Peach Belt      9 Lauren Posa    Jr    RS    6-1   ""      
#>  8 2024-2025 Augusta Peach Belt     10 Abigail Levin… Jr    OH    5-11  ""      
#>  9 2024-2025 Augusta Peach Belt     13 Jada Sugutura… Sr    MH    6-1   ""      
#> 10 2024-2025 Augusta Peach Belt     14 Libby NIckels  Fr    L/DS  5-1   ""      
#> # ℹ 71 more rows
#> # ℹ 20 more variables: `High School` <chr>, GP <dbl>, GS <dbl>, S <dbl>,
#> #   Kills <dbl>, Errors <dbl>, `Total Attacks` <dbl>, `Hit Pct` <dbl>,
#> #   Assists <dbl>, Aces <dbl>, SErr <dbl>, Digs <dbl>, RetAtt <dbl>,
#> #   RErr <dbl>, `Block Solos` <dbl>, `Block Assists` <dbl>, BErr <dbl>,
#> #   PTS <dbl>, BHE <dbl>, `Trpl Dbl` <dbl>
#> 
#> $teamdata
#> # A tibble: 5 × 20
#>   Season   Team  Conference     S Kills Errors `Total Attacks` `Hit Pct` Assists
#>   <chr>    <chr> <chr>      <dbl> <dbl>  <dbl>           <dbl>     <dbl>   <dbl>
#> 1 2024-20… Augu… Peach Belt   108  1374    618            4155     0.182    1258
#> 2 2024-20… Flag… Peach Belt   101  1243    477            3486     0.22     1152
#> 3 2024-20… Geor… Peach Belt   106  1056    572            3647     0.133     984
#> 4 2024-20… Land… Peach Belt   103  1288    448            3758     0.224    1162
#> 5 2024-20… USC … Peach Belt    78   984    453            2903     0.183     917
#> # ℹ 11 more variables: Aces <dbl>, SErr <dbl>, Digs <dbl>, RetAtt <dbl>,
#> #   RErr <dbl>, `Block Solos` <dbl>, `Block Assists` <dbl>, BErr <dbl>,
#> #   PTS <dbl>, BHE <dbl>, `Trpl Dbl` <dbl>
#> 
# }