Aggregate player statistics for a NCAA division and seasons

This is a wrapper around group_stats() that extracts season, match, or pbp data from players in all teams in the chosen division. For season stats, it aggregates all player data and team data into separate data frames and combines them into a list. For match and pbp stats, it aggregates into a data frame.

Usage

division_stats(
  year = NULL,
  division = 1,
  level = NULL,
  sport = "WVB",
  save = FALSE,
  path = "."
)

Arguments

year

Numeric vector of years for fall of desired seasons.

division

NCAA division (must be 1, 2, or 3).

level

Character string defining whether to aggregate "season", "match", or play-by-play ("pbp") data.

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

save

Logical for whether to save the statistics locally as CSVs (default FALSE).

path

Character string of path to save statistics files.

Value

For season level, returns list with data frames of player statistics and team statistics. For match and pbp levels, returns data frame of player statistics and play-by-play information respectively.

Note

This function requires internet connectivity as it checks the NCAA website for information.

See also