Match pattern to find team names

This is a convenience function to find NCAA team names in ncaa_teams. Once the proper team name is found, it can be passed to find_team_id() or group_stats().

Usage

find_team_name(pattern = NULL)

Arguments

pattern

Character string of pattern you want to find in the vector of team names.

Value

Returns a character vector of team names that include the submitted pattern.

See also

Examples

find_team_name(pattern = "Neb")
#> [1] "Neb. Wesleyan" "Neb.-Kearney"  "Nebraska"