Extract player statistics from a particular team and season

The NCAA's main page for a team includes a tab called "Team Statistics". This function extracts the table of player statistics for the season, as well as team and opponent statistics (though these can be omitted).

Usage

player_season_stats(team_id, team_stats = TRUE)

Arguments

team_id

Team ID determined by NCAA for season. To find ID, use find_team_id().

team_stats

Logical indicating whether to include (TRUE) or exclude (FALSE) team statistics. Default includes team statistics with player statistics.

Value

Returns a data frame of player statistics. Note that hometown and high school were added in 2024.

See also

Examples

# \donttest{
player_season_stats(team_id = "585290")
#> # A tibble: 16 × 29
#>    Season    Team     Conference Number Player        Yr    Pos   Ht    Hometown
#>    <chr>     <chr>    <chr>       <dbl> <chr>         <chr> <chr> <chr> <chr>   
#>  1 2024-2025 Nebraska Big Ten         2 Bergen Reilly So    S     6-1   Sioux F…
#>  2 2024-2025 Nebraska Big Ten         5 Rebekah Alli… Jr    MB    6-4   Lincoln…
#>  3 2024-2025 Nebraska Big Ten         6 Laney Choboy  So    L/DS  5-3   Raleigh…
#>  4 2024-2025 Nebraska Big Ten         7 Maisie Boesi… Jr    L/DS  5-6   Firth, …
#>  5 2024-2025 Nebraska Big Ten         8 Lexi Rodrigu… Sr    L/DS  5-5   Sterlin…
#>  6 2024-2025 Nebraska Big Ten         9 Kennedi Orr   Sr    S     6-0   Eagan, …
#>  7 2024-2025 Nebraska Big Ten        10 Olivia Mauch  Fr    L/DS  5-6   Benning…
#>  8 2024-2025 Nebraska Big Ten        11 Leyla Blackw… Sr    MB    6-4   San Die…
#>  9 2024-2025 Nebraska Big Ten        12 Taylor Landf… Sr    OH    6-5   Plainfi…
#> 10 2024-2025 Nebraska Big Ten        13 Merritt Beas… Sr    OH    6-4   Gardend…
#> 11 2024-2025 Nebraska Big Ten        15 Andi Jackson  So    MB    6-3   Brighto…
#> 12 2024-2025 Nebraska Big Ten        22 Lindsay Krau… Sr    OH    6-4   Papilli…
#> 13 2024-2025 Nebraska Big Ten        27 Harper Murray So    OH    6-2   Ann Arb…
#> 14 2024-2025 Nebraska Big Ten        NA TEAM          -     -     -     NA      
#> 15 2024-2025 Nebraska Big Ten        NA Totals        -     -     -     NA      
#> 16 2024-2025 Nebraska Big Ten        NA Opponent Tot… -     -     -     NA      
#> # ℹ 20 more variables: `High School` <chr>, GP <dbl>, GS <dbl>, S <dbl>,
#> #   Kills <dbl>, Errors <dbl>, `Total Attacks` <dbl>, `Hit Pct` <dbl>,
#> #   Assists <dbl>, Aces <dbl>, SErr <dbl>, Digs <dbl>, RetAtt <dbl>,
#> #   RErr <dbl>, `Block Solos` <dbl>, `Block Assists` <dbl>, BErr <dbl>,
#> #   PTS <dbl>, BHE <dbl>, `Trpl Dbl` <dbl>
# }