Extract arena, coach, record, and schedule information for a particular team and season

The NCAA's main page for a team includes a tab called "Schedule/Results". This function extracts information about the team's venue, coach, and records, as well as the table of the schedule and results. This returns a list, so you can subset specific components with $ (e.g., for coach information from an object called output, use output$coach).

Usage

team_season_info(team_id = NULL)

Arguments

team_id

Team ID determined by NCAA for season. To find ID, use find_team_id().

Value

Returns a list that includes arena, coach, schedule, and record information.

See also

Examples

# \donttest{
team_season_info(team_id = "585290")
#> $team_info
#>      team_id team_name conference_id conference div   yr    season
#> 1466  585290  Nebraska           827    Big Ten   1 2024 2024-2025
#> 
#> $arena
#>                  Arena name                    Capacity 
#> "Bob Devaney Sports Center"                     "7,907" 
#>                  Year built 
#>                      "1975" 
#> 
#> $coach
#>               Name         Alma mater            Seasons             Record 
#>        "John Cook" "San Diego - 1979"               "32"          "883-176" 
#> 
#> $record
#>      Overall record      Overall streak   Conference record   Conference streak 
#>      "33-3 (0.917)"        "Streak: L1"      "19-1 (0.950)"        "Streak: L1" 
#>         Home record         Home streak         Road record         Road streak 
#>      "22-0 (1.000)"       "Streak: W45"      "10-2 (0.833)"        "Streak: W1" 
#>      Neutral record      Neutral streak Non-division record Non-division streak 
#>       "1-1 (0.500)"        "Streak: L1"       "0-0 (0.000)"         "Streak: 0" 
#> 
#> $schedule
#> # A tibble: 36 × 4
#>    Date       Opponent                 Result Attendance
#>    <chr>      <chr>                    <chr>  <chr>     
#>  1 08/27/2024 Kentucky @Louisville, KY W 3-1  9,280     
#>  2 08/30/2024 A&M-Corpus Christi       W 3-0  8,956     
#>  3 08/31/2024 TCU                      W 3-1  8,695     
#>  4 09/03/2024 @ SMU                    L 0-3  6,773     
#>  5 09/05/2024 The Citadel              W 3-0  8,607     
#>  6 09/07/2024 Montana St.              W 3-0  8,456     
#>  7 09/10/2024 Creighton                W 3-2  8,924     
#>  8 09/13/2024 Arizona St.              W 3-0  8,772     
#>  9 09/14/2024 Wichita St.              W 3-0  8,541     
#> 10 09/18/2024 Stanford                 W 3-0  8,952     
#> # ℹ 26 more rows
#> 
# }