Extract teams statistics for season statistics from 2001-2024

The NCAA's main page for a team includes a tab called "Game By Game" and a section called "Career Totals". Though the page only shows one season's worth of information, this function extracts season summary stats starting with 2001. We have included the conference starting with 2020 (conference data for previous seasons is not currently available).

Usage

team_season_stats(team = NULL, opponent = FALSE, sport = "WVB")

Arguments

team

Name of school. Must match name used by NCAA. Find exact team name with find_team_name().

opponent

Logical indicating whether to include team's stats (FALSE) or opponent's stats (TRUE). Default is set to FALSE, returning team stats.

sport

Three letter abbreviation for NCAA sport (must be upper case; for example "WVB" for women's volleyball and "MVB" for men's volleyball).

Value

Returns a data frame of summary team statistics for each season.

See also

Examples

# \donttest{
team_season_stats(team = "Nebraska")
#>       Year     Team Conference   S Kills Errors Total Attacks Hit Pct Assists
#> 1  2001-02 Nebraska       <NA> 108  1790    532          4130   0.305    1575
#> 2  2002-03 Nebraska       <NA> 103  1747    481          3927   0.322    1532
#> 3  2003-04 Nebraska       <NA> 117  1915    720          4806   0.249    1748
#> 4  2004-05 Nebraska       <NA> 106  1823    610          4322   0.281    1657
#> 5  2005-06 Nebraska       <NA> 116  1986    581          4367   0.322    1823
#> 6  2006-07 Nebraska       <NA> 115  2001    637          4605   0.296    1863
#> 7  2007-08 Nebraska       <NA> 103  1829    506          4051   0.327    1691
#> 8  2008-09 Nebraska       <NA> 118  1714    590          4166   0.270    1578
#> 9  2009-10 Nebraska       <NA> 115  1685    618          4160   0.256    1568
#> 10 2010-11 Nebraska       <NA> 108  1609    528          3812   0.284    1501
#> 11 2011-12 Nebraska       <NA> 113  1532    504          3866   0.266    1426
#> 12 2012-13 Nebraska       <NA> 123  1728    509          4171   0.292    1617
#> 13 2013-14 Nebraska       <NA> 119  1710    593          4132   0.270    1597
#> 14 2014-15 Nebraska       <NA> 118  1564    599          4089   0.236    1476
#> 15 2015-16 Nebraska       <NA> 129  1873    589          4690   0.274    1730
#> 16 2016-17 Nebraska       <NA> 117  1688    533          4026   0.287    1558
#> 17 2017-18 Nebraska       <NA> 125  1777    575          4267   0.282    1632
#> 18 2018-19 Nebraska       <NA> 133  1772    582          4491   0.265    1651
#> 19 2019-20 Nebraska       <NA> 119  1641    567          4029   0.267    1526
#> 20 2020-21 Nebraska       <NA>  66   925    323          2229   0.270     847
#> 21 2021-22 Nebraska       <NA> 121  1677    667          4551   0.222    1548
#> 22 2022-23 Nebraska       <NA> 109  1461    519          3866   0.244    1331
#> 23 2023-24 Nebraska       <NA> 121  1676    597          3959   0.273    1553
#> 24 2024-25 Nebraska       <NA> 122  1721    551          4114   0.284    1596
#>    Aces SErr Digs RetAtt RErr Block Solos Block Assists BErr    PTS BHE
#> 1   162   NA 1833     NA   NA         116           670   NA     NA  NA
#> 2   169   NA 1642     NA   NA          81           626   NA     NA  NA
#> 3   225   NA 1941     NA   NA          80           582   NA     NA  NA
#> 4   166   NA 1846     NA   NA          89           675   NA     NA  NA
#> 5   162   NA 1731     NA   NA          66           823   NA     NA  NA
#> 6   175   NA 1942     NA   NA          94           531   NA     NA  NA
#> 7   187   NA 1687     NA   NA          90           397   NA     NA  NA
#> 8   154   NA 1734     NA   NA          78           451   NA     NA  NA
#> 9   130   NA 1837     NA   NA          75           398   NA     NA  NA
#> 10  127  148 1698     NA   77          73           510   79 2064.0  11
#> 11   91  161 1696     NA   95          52           484   77 1917.0  18
#> 12  139  217 1767     NA   94          79           477   68 2184.5  15
#> 13  115  241 1714     NA   82          69           538   96 2163.0  20
#> 14  118  220 1699     NA   93          72           543   64 2025.5  23
#> 15  140  232 2049     NA   79          55           605   67 2370.5  24
#> 16  142  215 1785     NA   69          57           546   59 2160.0  19
#> 17  184  236 1904     NA   80          61           479   68 2261.5  22
#> 18  215  285 2100     NA  124          56           601   99 2343.5  14
#> 19  125  238 1802   2087  126          50           470   66 1940.0  11
#> 20  104  131  957   1155   71          21           269   40 1184.5  12
#> 21  181  241 2054   2164  105          38           530   54 2161.0   8
#> 22  128  261 1721   1824   77          33           537   47 1890.5  25
#> 23  143  309 1666   2018  104          39           574   43 2145.0  15
#> 24  155  193 1856   1920   81          51           581   50 2217.5   1
#>    Trpl Dbl
#> 1         1
#> 2         1
#> 3         1
#> 4         1
#> 5         1
#> 6         1
#> 7         1
#> 8         1
#> 9         1
#> 10       NA
#> 11       NA
#> 12       NA
#> 13       NA
#> 14       NA
#> 15       NA
#> 16       NA
#> 17       NA
#> 18       NA
#> 19       10
#> 20       19
#> 21       34
#> 22       32
#> 23       35
#> 24       36
# }